+++ Assets/Scripts/Utils/IHealth.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IHealth 
 {
-    float Health { get; set; }
-    void TakeDamage(float amount);
+    int Health { get; set; }
+    void TakeDamage(int amount);
     void Die();
 }
